fix: Path iterator returned wrong node
Enabled all proof tests, which discovered an introduced bug when
MaybePersisted was introduced. The node being returned was the wrong
one, which caused some proof errors when constructing proofs, notably
the wrong node was being returned if the node wasn't yet persisted.

Since the tests run with unpersisted but hashed nodes, this caused
some failures when the tests were re-enabled.

Also restructured the tests. Proof tests are separated now from
range proof tests.

Range proof tests were moved but not enabled in this PR (will be
done next).
