+++ firewood/src/merkle/tests/range.rs
@@ -7,237 +7,6 @@ use crate::range_proof::RangeProof;
 type KeyValuePairs = Vec<(Box<[u8]>, Box<[u8]>)>;
 
 #[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn range_proof_invalid_bounds() {
-    let merkle = create_in_memory_merkle().hash();
-
-    let start_key = &[0x01];
-    let end_key = &[0x00];
-
-    match merkle.range_proof(Some(start_key), Some(end_key), NonZeroUsize::new(1)) {
-        Err(api::Error::InvalidRange {
-            start_key: first_key,
-            end_key: last_key,
-        }) if *first_key == *start_key && *last_key == *end_key => (),
-        Err(api::Error::InvalidRange { .. }) => panic!("wrong bounds on InvalidRange error"),
-        _ => panic!("expected InvalidRange error"),
-    }
-}
-
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn full_range_proof() {
-    let merkle = init_merkle((u8::MIN..=u8::MAX).map(|k| ([k], [k])));
-
-    let rangeproof = merkle.range_proof(None, None, None).unwrap();
-    assert_eq!(rangeproof.key_values().len(), u8::MAX as usize + 1);
-    assert_ne!(rangeproof.start_proof(), rangeproof.end_proof());
-    let left_proof = merkle.prove(&[u8::MIN]).unwrap();
-    let right_proof = merkle.prove(&[u8::MAX]).unwrap();
-    assert_eq!(rangeproof.start_proof(), &left_proof);
-    assert_eq!(rangeproof.end_proof(), &right_proof);
-}
-
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn single_value_range_proof() {
-    const RANDOM_KEY: u8 = 42;
-
-    let merkle = init_merkle((u8::MIN..=u8::MAX).map(|k| ([k], [k])));
-
-    let rangeproof = merkle
-        .range_proof(Some(&[RANDOM_KEY]), None, NonZeroUsize::new(1))
-        .unwrap();
-    assert_eq!(rangeproof.start_proof(), rangeproof.end_proof());
-    assert_eq!(rangeproof.key_values().len(), 1);
-}
-
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn shared_path_proof() {
-    let key1 = b"key1";
-    let value1 = b"1";
-
-    let key2 = b"key2";
-    let value2 = b"2";
-
-    let merkle = init_merkle([(key1, value1), (key2, value2)]);
-
-    let root_hash = merkle.nodestore().root_hash().unwrap();
-
-    let key = key1;
-    let proof = merkle.prove(key).unwrap();
-    proof.verify(key, Some(value1), &root_hash).unwrap();
-
-    let key = key2;
-    let proof = merkle.prove(key).unwrap();
-    proof.verify(key, Some(value2), &root_hash).unwrap();
-}
-
-// this was a specific failing case
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn shared_path_on_insert() {
-    init_merkle([
-        (
-            &[1, 1, 46, 82, 67, 218][..],
-            &[23, 252, 128, 144, 235, 202, 124, 243][..],
-        ),
-        (
-            &[1, 0, 0, 1, 1, 0, 63, 80],
-            &[99, 82, 31, 213, 180, 196, 49, 242],
-        ),
-        (
-            &[0, 0, 0, 169, 176, 15],
-            &[105, 211, 176, 51, 231, 182, 74, 207],
-        ),
-        (
-            &[1, 0, 0, 0, 53, 57, 93],
-            &[234, 139, 214, 220, 172, 38, 168, 164],
-        ),
-    ]);
-}
-
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn overwrite_leaf() {
-    let key = &[0x00];
-    let val = &[1];
-    let overwrite = &[2];
-
-    let mut merkle = create_in_memory_merkle();
-
-    merkle.insert(key, val[..].into()).unwrap();
-
-    assert_eq!(
-        merkle.get_value(key).unwrap().as_deref(),
-        Some(val.as_slice())
-    );
-
-    merkle.insert(key, overwrite[..].into()).unwrap();
-
-    assert_eq!(
-        merkle.get_value(key).unwrap().as_deref(),
-        Some(overwrite.as_slice())
-    );
-}
-
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn single_key_proof_with_one_node() {
-    let key = b"key";
-    let value = b"value";
-
-    let merkle = init_merkle([(key, value)]);
-
-    let root_hash = merkle.nodestore().root_hash().unwrap();
-
-    let proof = merkle.prove(key).unwrap();
-    proof.verify(key, Some(value), &root_hash).unwrap();
-}
-
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn two_key_proof_without_shared_path() {
-    let key1 = &[0x00];
-    let key2 = &[0xff];
-
-    let merkle = init_merkle([(key1, key1), (key2, key2)]);
-
-    let root_hash = merkle.nodestore().root_hash().unwrap();
-
-    let proof = merkle.prove(key1).unwrap();
-    proof.verify(key1, Some(key1), &root_hash).unwrap();
-
-    let proof = merkle.prove(key2).unwrap();
-    proof.verify(key2, Some(key2), &root_hash).unwrap();
-}
-
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn test_proof() {
-    let rng = firewood_storage::SeededRng::from_env_or_random();
-    let set = fixed_and_pseudorandom_data(&rng, 500);
-    let mut items = set.iter().collect::<Vec<_>>();
-    items.sort_unstable();
-    let merkle = init_merkle(items.clone());
-
-    let root_hash = merkle.nodestore().root_hash().unwrap();
-
-    for (key, val) in items {
-        let proof = merkle.prove(key).unwrap();
-        assert!(!proof.is_empty());
-        proof.verify(key, Some(val), &root_hash).unwrap();
-    }
-}
-
-#[test]
-/// Verify the proofs that end with leaf node with the given key.
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn test_proof_end_with_leaf() {
-    let merkle = init_merkle([
-        ("do", "verb"),
-        ("doe", "reindeer"),
-        ("dog", "puppy"),
-        ("doge", "coin"),
-        ("horse", "stallion"),
-        ("ddd", "ok"),
-    ]);
-    let root_hash = merkle.nodestore().root_hash().unwrap();
-
-    let key = b"doe";
-
-    let proof = merkle.prove(key).unwrap();
-    assert!(!proof.is_empty());
-
-    proof.verify(key, Some(b"reindeer"), &root_hash).unwrap();
-}
-
-#[test]
-/// Verify the proofs that end with branch node with the given key.
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn test_proof_end_with_branch() {
-    let items = [
-        ("d", "verb"),
-        ("do", "verb"),
-        ("doe", "reindeer"),
-        ("e", "coin"),
-    ];
-    let merkle = init_merkle(items);
-    let root_hash = merkle.nodestore().root_hash().unwrap();
-
-    let key = b"d";
-
-    let proof = merkle.prove(key).unwrap();
-    assert!(!proof.is_empty());
-
-    proof.verify(key, Some(b"verb"), &root_hash).unwrap();
-}
-
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn test_bad_proof() {
-    let rng = firewood_storage::SeededRng::from_env_or_random();
-    let set = fixed_and_pseudorandom_data(&rng, 800);
-    let mut items = set.iter().collect::<Vec<_>>();
-    items.sort_unstable();
-    let merkle = init_merkle(items.clone());
-    let root_hash = merkle.nodestore().root_hash().unwrap();
-
-    for (key, value) in items {
-        let proof = merkle.prove(key).unwrap();
-        assert!(!proof.is_empty());
-
-        // Delete an entry from the generated proofs.
-        let mut new_proof = proof.into_mutable();
-        new_proof.pop();
-
-        // TODO: verify error result matches expected error
-        assert!(new_proof.verify(key, Some(value), &root_hash).is_err());
-    }
-}
-
-#[test]
 // Tests that missing keys can also be proven. The test explicitly uses a single
 // entry trie and checks for missing keys both before and after the single entry.
 #[ignore = "https://github.com/ava-labs/firewood/issues/738"]
@@ -253,17 +22,6 @@ fn test_missing_key_proof() {
 
         proof.verify(key, None::<&[u8]>, &root_hash).unwrap();
     }
-}
-
-#[test]
-#[ignore = "https://github.com/ava-labs/firewood/issues/738"]
-fn test_empty_tree_proof() {
-    let items: Vec<(&str, &str)> = Vec::new();
-    let merkle = init_merkle(items);
-    let key = "x".as_ref();
-
-    let proof_err = merkle.prove(key).unwrap_err();
-    assert!(matches!(proof_err, ProofError::Empty), "{proof_err:?}");
 }
 
 #[test]
