+++ firewood/src/proof.rs
@@ -74,7 +74,7 @@ pub enum ProofError {
     EmptyRange,
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Default)]
+#[derive(Clone, PartialEq, Eq, Default)]
 /// A node in a proof.
 pub struct ProofNode {
     /// The key this node is at. Each byte is a nibble.
@@ -87,6 +87,33 @@ pub struct ProofNode {
     pub value_digest: Option<ValueDigest<Value>>,
     /// The hash of each child, or None if the child does not exist.
     pub child_hashes: Children<HashType>,
+}
+
+impl std::fmt::Debug for ProofNode {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let key: Vec<u8> = self.key().collect();
+        let children: Vec<(usize, TrieHash)> = self
+            .children()
+            .iter()
+            .enumerate()
+            .filter_map(|(i, c)| c.as_ref().map(|h| (i, (*h).clone().into_triehash())))
+            .collect();
+        let value_digest = self.value_digest();
+
+        let mut ds = f.debug_struct("ProofNode");
+        ds.field("key_nibbles", &key)
+            .field("value_digest", &value_digest)
+            .field("children", &children);
+
+        #[cfg(feature = "ethhash")]
+        ds.field("partial_len", &self.partial_len);
+
+        // Compute the hash and render it as well
+        let hash = firewood_storage::Preimage::to_hash(self);
+        ds.field("hash", &hash);
+
+        ds.finish()
+    }
 }
 
 impl Hashable for ProofNode {
